+++ packages/react-story-rich-ui/src/Layout/Layout.md
@@ -10,43 +10,67 @@ import Grid from '@material-ui/core/Grid';
 import Button from '@material-ui/core/Button';
 
 import Layout from '@react-story-rich/ui/Layout';
-import { Provider, Story, Element, goForward, goTo } from '@react-story-rich/core';
+import { Provider, Story, Element } from '@react-story-rich/core';
 
-// forwardRef is important otherwise component will not be auto focused (+ throwing a warning)
-const CustomElement = forwardRef(({ children, forwardProps, ...rest }, ref) =>  (
-  <Grid sm={4} item>
-    <Card ref={ref} {...rest}>
-      <CardContent>
-        <Typography variant="body2" color="textSecondary" component="p">
-          {children}
-        </Typography>
-      </CardContent>
-    </Card>
-  </Grid>
-));
+class CustomElement extends Element {
+  render() {
+    const { _id, autoFocus, children, enabled, label, tabIndex } = this.props;
+console.log(this);
+
+    return (
+      <Grid sm={4} item>
+        <Card
+          autoFocus={autoFocus}
+          data-id={_id}
+          data-index={this.index}
+          data-label={label}
+          data-disabled={!enabled}
+          onClick={this._handleTap}
+          onKeyPress={this._handleKeyPress}
+          ref={this.ref}
+          tabIndex={tabIndex}
+        >
+          <CardContent>
+            <Typography variant="body2" color="textSecondary">
+              {children}
+            </Typography>
+          </CardContent>
+        </Card>
+      </Grid>
+    ) 
+  }
+}
+
 
 CustomElement.propTypes = {
   children: PropTypes.node.isRequired,
 };
 
-function SmallSequence() {
+const location = 3;
+const history = [
+  { from: 0, to: 1 },
+  { from: 1, to: 3 },
+];
+
+
+const Knot = () => {
   const timeout = 1000;
   let count = 0;
 
-  const onTimeout = ({ dispatch }) => dispatch(goForward());
-  const loop = ({ dispatch, location }) => {
+  const onTimeout = () => { this.goForward() };
+  const loop = () => {
     count += 1;
-    if (count < 3) { dispatch(goTo(location, 0)); }
+    if (count < 3) { this.goTo(() => 0); }
   };
 
   return (
     <>
-      <Element component={CustomElement} timeout={timeout} onTimeout={onTimeout}>Hey</Element>
-      <Element component={CustomElement} timeout={timeout} onTimeout={onTimeout}>Oh !</Element>
-      <Element component={CustomElement} onTap={loop}>Let&#39;s go !</Element>
+      <CustomElement timeout={timeout} onTimeout={onTimeout}>Hey</CustomElement>
+      <CustomElement timeout={timeout} onTimeout={onTimeout}>Oh !</CustomElement>
+      <CustomElement onTap={loop}>Let&#39;s go !</CustomElement>
     </>
   );
-}
+};
 
 const [start, setStart] = useState(false);
 const handleClick = useCallback(() => setStart(true), [setStart]);
@@ -56,8 +80,13 @@ const handleClick = useCallback(() => setStart(true), [setStart]);
       <Container>
         {!start && <Button variant="contained" color="primary" onClick={handleClick}>Start</Button>}
         {start && (
-          <Story component={Grid} container spacing={2} scrollToBottom={false}>
-            {SmallSequence()}
+          <Story
+            component={Grid}
+            componentProps={{ container: true, spacing: 2 }}
+            autoScroll={false}
+            autoFocus={false}
+          >
+            {Knot()}
           </Story>
         )}
       </Container>
