+++ packages/react-story-rich/src/components/Knot.js
