+++ packages/react-story-rich/src/components/Navigation.js
