+++ packages/react-story-rich/src/containers/mapStateToProps.js
@@ -1,5 +1,4 @@
 const mapStateToProps = (state) => ({
-  data: state.data,
   history: state.history,
   location: state.location,
 });
