+++ packages/react-story-rich/src/index.js
@@ -1,3 +1,2 @@
 export * from './containers';
-export * from './store/data';
 export * from './store/history';
