--- packages/react-story-rich/src/store/data/index.js
@@ -1,14 +1,0 @@
-import createReducer from '../createReducer';
-
-// ACTIONS
-export const SET_DATA = '@react-story-rich.SET_DATA';
-export const setData = (data) => ({ type: SET_DATA, data });
-
-export const UPDATE_DATA = '@react-story-rich.UPDATE_DATA';
-export const updateData = (data) => ({ type: UPDATE_DATA, data });
-
-// REDUCER
-export default createReducer({}, {
-  [SET_DATA]: (state, { data }) => data,
-  [UPDATE_DATA]: (state, { data }) => ({ ...state, ...data }),
-});
