+++ packages/react-story-rich/src/store/location/index.js
@@ -1,10 +1,11 @@
 import createReducer from '../createReducer';
-import { GO_BACKWARD_TO, GO_TO, SET_LOCATION } from '../history';
+import { SET_LOCATION } from '../history';
+import { GO_TO, REWIND_TO } from '../../components/Navigation';
 
 const set = (state, { location, to }) => (location !== undefined ? location : to);
 
 export default createReducer(0, {
-  [GO_BACKWARD_TO]: set,
+  [REWIND_TO]: set,
   [GO_TO]: set,
   [SET_LOCATION]: set,
 });
